backend/requirements.txt
Flask
requests
openai
sqlalchemy
python-dotenv
werkzeug
Flask-Cors
pymongo
python-dateutil
Flask-Session